pyproject.toml
[project]
name = "nba_etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "dagster-cloud",
    "pandas",
    "sqlalchemy",
    "nba_api",
    "python-dotenv",
    "openai<1.0.0",  # compatible with current embedding code path
    "pymongo",
    # optional vector DB for later
    "qdrant-client~=1.10",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "nba_etl.definitions"
code_location_name = "nba_etl"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["nba_etl_tests"]
